pyproject.toml
[project]
name = "whatbytes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
